Refactor: python, cts: Move Reattach into the _cts module.
No code changes.
